pyproject.toml
[project]
name = "udelar-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipdb>=0.13.13",
    "loguru>=0.7.3",
    "neo4j>=5.28.1",
    "polars>=1.30.0",
    "python-levenshtein>=0.27.1",
    "typer>=0.15.4",
    "unidecode>=1.4.0",
]

[project.scripts]
udegraph = "udelar_graph.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
select = ["F", "I", "E", "W", "ARG", "T201"]
fixable = ["I"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E712", "ARG001"]
"research/*.py" = ["T201", "ARG001"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.mypy]
ignore_missing_imports = true
show_error_codes = true


[[tool.mypy.overrides]]
module = ["torch.*"]
follow_imports = "skip"
follow_imports_for_stubs = true

[dependency-groups]
crawlers = [
    "beautifulsoup4>=4.13.4",
    "scrapy>=2.13.0",
]
dev = [
    "dvc>=3.59.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "mypy>=1.15.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.11",
]
